Screen scraper update (#303)
* Fix screen-scraping character encoding

Fix mangling of multi-byte utf-8 encoded characters.

Add a test for the "charset=" <meta/> information so that it works
for both HTML 4 and HTML 5. Previous code only worked for HTML 4.

Use "in" for the search instead of a regular expression, because
the match string doesn't contain RE operators.

When processing a page that's already in utf-8, encode the translation
of named HTML entities into utf-8 instead of trying to put unicode
characters into the 8-bit encoded HTML string.

If the page is already utf-8 encoded, don't re-encode it into utf-8.

* Fix screen-scraping of general info

Add utf-8 decodings of some HTML entities to allow them to be edited out of
the text.

Update the general info and genre screenscraping regular expressions
and the code that cleans up the matched data for both display and
text file output.

In matching the genre, use re.search() instead of re.matchall(),
because only the first match is used anyway.

Generalise the htmltags regular expression to allow matching of
multi-line HTML tags.

In the scraped data affected by these changes, clean up spacing using
' '.join(<string>.split()
rather than the stripmask regular expression. It has the advantage
of also doing the equivalent to a <string>.strip() as well.

Remove substitution of HTML <br> from the "Country" data in
IMDb.IMDBsavetxt(), because the country data no longer seems to use
them.

* Fix screen-scraping of cast list

Update the cast list screenscraping regular expression
and the code that cleans up the matched data for both display and
text file output.

Use utf-8 decodings of some HTML entities to allow them to be edited out of
the text.

In the scraped data affected by these changes, clean up spacing using
' '.join(<string>.split()
rather than the stripmask regular expression. It has the advantage
of also doing the equicalent to a <string>.strip() as well.

Removed processing of unused "additional" match fields in the cast list.

* Fix screen-scraping of general info

Add utf-8 decodings of some HTML entities to allow them to be edited
out of the text.

Update the extra info regular expressions and the code that cleans
up the matched data for both display and text file output.

Split out the awards regular expression, because it needs to be
handled differently from the other fields.

In the scraped data affected by these changes, clean up spacing
using
' '.join(<string>.split()
rather than the stripmask regular expression. It has the advantage
of also doing the equivalent to a <string>.strip() as well.

Add additional spacing to the extrainfo display to improve readability.

* Fix screen-scraping of general info

Add utf-8 decodings of some HTML entities to allow them to be edited
out of the text.

Update the extra info regular expressions and the code that cleans
up the matched data for both display and text file output.

Split out the awards regular expression, because it needs to be
handled differently from the other fields.

In the scraped data affected by these changes, clean up spacing
using
' '.join(<string>.split()
rather than the stripmask regular expression. It has the advantage
of also doing the equivalent to a <string>.strip() as well.

Add additional spacing to the extrainfo display to improve readability.

* Optionally display full match title info in title menu screen

* Add help to the main ActionMap in the IMDB class

Split the ActionMap into several HelpableActionMaps so that it works
on systems that don't support ActionMap with lists of contexts.

* Translate some untranslated UI strings